fix ts declare global
