+++ game/scripts/vscripts/api/global.lua
@@ -1,0 +1,5 @@
+local ____lualib = require("lualib_bundle")
+local __TS__SourceMapTraceBack = ____lualib.__TS__SourceMapTraceBack
+__TS__SourceMapTraceBack(debug.getinfo(1).short_src, {});
+local ____exports = {}
+return ____exports
