+++ game/scripts/vscripts/api/player.lua
@@ -3,14 +3,15 @@ local __TS__Class = ____lualib.__TS__Class
 local __TS__New = ____lualib.__TS__New
 local __TS__ArrayFind = ____lualib.__TS__ArrayFind
 local __TS__SourceMapTraceBack = ____lualib.__TS__SourceMapTraceBack
-__TS__SourceMapTraceBack(debug.getinfo(1).short_src, {["8"] = 2,["9"] = 2,["10"] = 3,["11"] = 3,["12"] = 3,["13"] = 5,["14"] = 5,["16"] = 5,["17"] = 11,["18"] = 11,["19"] = 11,["21"] = 11,["22"] = 17,["23"] = 17,["24"] = 17,["26"] = 17,["27"] = 36,["28"] = 36,["30"] = 36,["31"] = 56,["32"] = 56,["33"] = 56,["35"] = 60,["36"] = 61,["37"] = 62,["38"] = 62,["39"] = 62,["40"] = 62,["41"] = 62,["42"] = 62,["43"] = 62,["44"] = 62,["45"] = 62,["46"] = 62,["47"] = 62,["48"] = 62,["49"] = 65,["50"] = 66,["51"] = 66,["54"] = 59,["55"] = 75,["56"] = 76,["57"] = 78,["59"] = 79,["60"] = 79,["61"] = 80,["62"] = 81,["64"] = 79,["67"] = 84,["68"] = 85,["69"] = 85,["70"] = 85,["71"] = 88,["72"] = 88,["73"] = 88,["74"] = 85,["75"] = 85,["76"] = 85,["77"] = 85,["78"] = 85,["79"] = 95,["80"] = 97,["81"] = 75,["82"] = 100,["83"] = 101,["84"] = 102,["85"] = 103,["86"] = 104,["87"] = 105,["88"] = 108,["89"] = 109,["90"] = 111,["91"] = 112,["92"] = 100,["93"] = 115,["94"] = 116,["95"] = 117,["97"] = 119,["98"] = 115,["99"] = 123,["100"] = 124,["101"] = 125,["104"] = 129,["105"] = 130,["106"] = 130,["107"] = 130,["108"] = 130,["109"] = 132,["110"] = 132,["111"] = 132,["113"] = 132,["116"] = 136,["117"] = 137,["119"] = 123,["120"] = 141,["122"] = 142,["123"] = 142,["124"] = 143,["125"] = 145,["126"] = 146,["127"] = 146,["128"] = 146,["129"] = 146,["130"] = 147,["131"] = 149,["132"] = 149,["133"] = 149,["134"] = 149,["135"] = 149,["138"] = 142,["141"] = 141,["142"] = 155,["144"] = 156,["145"] = 156,["146"] = 157,["147"] = 159,["148"] = 160,["149"] = 160,["150"] = 160,["151"] = 160,["152"] = 161,["153"] = 163,["154"] = 163,["155"] = 163,["156"] = 163,["157"] = 163,["160"] = 156,["163"] = 155,["164"] = 168,["165"] = 169,["166"] = 169,["167"] = 169,["168"] = 169,["169"] = 170,["170"] = 171,["172"] = 173,["173"] = 168,["174"] = 176,["175"] = 177,["176"] = 177,["177"] = 177,["178"] = 177,["179"] = 178,["180"] = 179,["182"] = 181,["183"] = 176,["184"] = 184,["185"] = 185,["186"] = 185,["187"] = 185,["188"] = 185,["189"] = 184,["190"] = 189,["191"] = 190,["192"] = 191,["193"] = 193,["194"] = 193,["195"] = 193,["196"] = 193,["197"] = 193,["198"] = 193,["199"] = 193,["200"] = 205,["201"] = 189,["202"] = 208,["203"] = 209,["204"] = 210,["205"] = 211,["206"] = 213,["207"] = 216,["208"] = 216,["209"] = 216,["210"] = 216,["211"] = 217,["212"] = 218,["213"] = 219,["215"] = 221,["216"] = 221,["217"] = 221,["218"] = 221,["219"] = 222,["220"] = 223,["222"] = 225,["223"] = 225,["225"] = 228,["227"] = 208,["228"] = 232,["229"] = 233,["230"] = 234,["231"] = 232,["232"] = 57,["233"] = 58});
+__TS__SourceMapTraceBack(debug.getinfo(1).short_src, {["8"] = 2,["9"] = 2,["10"] = 3,["11"] = 3,["12"] = 3,["13"] = 5,["14"] = 5,["15"] = 5,["17"] = 5,["18"] = 11,["19"] = 11,["20"] = 11,["22"] = 11,["23"] = 17,["24"] = 17,["25"] = 17,["27"] = 17,["28"] = 36,["29"] = 36,["31"] = 36,["32"] = 42,["33"] = 42,["34"] = 42,["36"] = 46,["37"] = 47,["38"] = 48,["39"] = 48,["40"] = 48,["41"] = 48,["42"] = 48,["43"] = 48,["44"] = 48,["45"] = 48,["46"] = 48,["47"] = 48,["48"] = 48,["49"] = 48,["50"] = 51,["51"] = 52,["52"] = 52,["55"] = 45,["56"] = 61,["57"] = 62,["58"] = 64,["60"] = 65,["61"] = 65,["62"] = 66,["63"] = 67,["65"] = 65,["68"] = 70,["69"] = 71,["70"] = 71,["71"] = 71,["72"] = 74,["73"] = 74,["74"] = 74,["75"] = 71,["76"] = 71,["77"] = 71,["78"] = 71,["79"] = 71,["80"] = 81,["81"] = 83,["82"] = 61,["83"] = 86,["84"] = 87,["85"] = 88,["86"] = 89,["87"] = 90,["88"] = 91,["89"] = 94,["90"] = 95,["91"] = 97,["92"] = 98,["93"] = 86,["94"] = 101,["95"] = 102,["96"] = 103,["98"] = 105,["99"] = 101,["100"] = 109,["101"] = 110,["102"] = 111,["105"] = 115,["106"] = 116,["107"] = 116,["108"] = 116,["109"] = 116,["110"] = 118,["111"] = 118,["112"] = 118,["114"] = 118,["117"] = 122,["118"] = 123,["120"] = 109,["121"] = 127,["123"] = 128,["124"] = 128,["125"] = 129,["126"] = 131,["127"] = 132,["128"] = 132,["129"] = 132,["130"] = 132,["131"] = 133,["132"] = 135,["133"] = 135,["134"] = 135,["135"] = 135,["136"] = 135,["139"] = 128,["142"] = 127,["143"] = 141,["145"] = 142,["146"] = 142,["147"] = 143,["148"] = 145,["149"] = 146,["150"] = 146,["151"] = 146,["152"] = 146,["153"] = 147,["154"] = 149,["155"] = 149,["156"] = 149,["157"] = 149,["158"] = 149,["161"] = 142,["164"] = 141,["165"] = 154,["166"] = 155,["167"] = 155,["168"] = 155,["169"] = 155,["170"] = 156,["171"] = 157,["173"] = 159,["174"] = 154,["175"] = 162,["176"] = 163,["177"] = 163,["178"] = 163,["179"] = 163,["180"] = 164,["181"] = 165,["183"] = 167,["184"] = 162,["185"] = 170,["186"] = 171,["187"] = 171,["188"] = 171,["189"] = 171,["190"] = 170,["191"] = 175,["192"] = 176,["193"] = 177,["194"] = 179,["195"] = 179,["196"] = 179,["197"] = 179,["198"] = 179,["199"] = 179,["200"] = 179,["201"] = 191,["202"] = 175,["203"] = 194,["204"] = 195,["205"] = 196,["206"] = 197,["207"] = 199,["208"] = 202,["209"] = 202,["210"] = 202,["211"] = 202,["212"] = 203,["213"] = 204,["214"] = 205,["216"] = 207,["217"] = 207,["218"] = 207,["219"] = 207,["220"] = 208,["221"] = 209,["223"] = 211,["224"] = 211,["226"] = 214,["228"] = 194,["229"] = 218,["230"] = 219,["231"] = 220,["232"] = 218,["233"] = 43,["234"] = 44});
 local ____exports = {}
 local ____property_controller = require("modifiers.property.property_controller")
 local PropertyController = ____property_controller.PropertyController
 local ____api_client = require("api.api_client")
 local ApiClient = ____api_client.ApiClient
 local HttpMethod = ____api_client.HttpMethod
-local MemberDto = __TS__Class()
+____exports.MemberDto = __TS__Class()
+local MemberDto = ____exports.MemberDto
 MemberDto.name = "MemberDto"
 function MemberDto.prototype.____constructor(self)
 end
